requirements.txt
fastapi
sqlmodel
pydantic_settings
